requirements.txt
selenium
webdriver-manager
chromedriver-binary
python-dotenv
timeunit